Moved functions to a better named module.
