+++ satkit/export/__init__.py
@@ -32,11 +32,10 @@
 """Export functions for extracting parts of the data."""
 
 from .images import (
+    export_aggregate_image_and_meta, export_aggregate_image_meta,
+    export_ultrasound_frame_and_meta,
+    export_ultrasound_frame_meta,
     publish_distance_matrix,
     publish_aggregate_images
 )
 
-from .export import (
-    export_aggregate_image_meta, export_aggregate_image_and_meta,
-    export_ultrasound_frame_meta, export_ultrasound_frame_and_meta
-)
