+++ satkit/export/export.py
@@ -31,174 +31,4 @@
 #
 """Export functions."""
 
-import logging
-from pathlib import Path
-
-import nestedtext
-from matplotlib.figure import Figure
-from PIL import Image
-
-from satkit.data_structures import Recording, Session
-from satkit.metrics import AggregateImage, AggregateImageParameters
-from satkit.save_and_load import nested_text_converters
-
-_logger = logging.getLogger('satkit._saver')
-
-
-def export_ultrasound_frame_meta(
-        filename: str | Path,
-        session: Session,
-        recording: Recording,
-        selection_index: int,
-        selection_time: float,
-) -> None:
-    """
-    Write ultrasound frame metadata to a human-readable text file.
-
-    The purpose of this function is to generate a file documenting an extracted
-    ultrasound frame, so that it can be found again in its original context.
-
-    Parameters
-    ----------
-    filename : str | Path
-        Filename or path of the extracted ultrasound frame.
-    session : Session
-        Session that the frame belongs to.
-    recording : Recording
-        Recording that the frame belongs to.
-    selection_index : int
-        Index of the frame within the ultrasound video.
-    selection_time : float
-        Time in seconds of the frame within the **recording**. This is relative
-        to what ever -- most likely the beginning of audio -- is being used as
-        t=0s.
-    """
-    if not isinstance(filename, Path):
-        filename = Path(filename)
-    meta_filename = filename.with_suffix('.txt')
-    with meta_filename.open('w', encoding='utf-8') as file:
-        file.write(f"Metadata for frame extracted by SATKIT to {filename}.\n")
-        file.write(f"Session path: {session.recorded_path}\n")
-        file.write(f"Participant ID: {recording.meta_data.participant_id}\n")
-        file.write(f"Recording filename: {recording.name}\n")
-        file.write(f"Recorded at: {recording.meta_data.time_of_recording}\n")
-        file.write(f"Prompt: {recording.meta_data.prompt}\n")
-        file.write(f"Frame number: {selection_index}\n")
-        file.write(f"Timestamp in recording: {selection_time}\n")
-        _logger.debug("Wrote file %s.", meta_filename)
-
-
-def export_ultrasound_frame_and_meta(
-        filename: str | Path,
-        figure: Figure,
-        session: Session,
-        recording: Recording,
-        selection_index: int,
-        selection_time: float,
-        bbox_inches: str = 'tight',
-        pad_inches: float = 0.05,
-) -> None:
-    """
-    Write ultrasound frame metadata to a human-readable text file.
-
-    The purpose of this function is to generate a file documenting an extracted
-    ultrasound frame, so that it can be found again in its original context.
-
-    Parameters
-    ----------
-    figure : matplotlib.figure.Figure
-        Figure to write in the image file.
-    filename : str | Path
-        Filename or path for the ultrasound frame to export. Format is deduced
-        from the file suffix.
-    session : Session
-        Session that the frame belongs to.
-    recording : Recording
-        Recording that the frame belongs to.
-    selection_index : int
-        Index of the frame within the ultrasound video.
-    selection_time : float
-        Time in seconds of the frame within the **recording**. This is relative
-        to what ever -- most likely the beginning of audio -- is being used as
-        t=0s.
-    bbox_inches : str
-        bounding box inches for the figure, by default 'tight'.
-    pad_inches :
-        padding inches around the figure, by default 0.05.
-    """
-    figure.savefig(filename, bbox_inches=bbox_inches, pad_inches=pad_inches)
-    _logger.debug("Wrote file %s.", filename)
-
-    export_ultrasound_frame_meta(
-        filename=filename,
-        session=session,
-        recording=recording,
-        selection_index=selection_index,
-        selection_time=selection_time,
-    )
-
-
-def export_aggregate_image_meta(
-        filename: str | Path,
-        session: Session,
-        recording: Recording,
-        aggregate_meta: AggregateImageParameters,
-) -> None:
-    """
-    Write ultrasound frame metadata to a human-readable text file.
-
-    The purpose of this function is to generate a file documenting an extracted
-    ultrasound frame, so that it can be found again in its original context.
-
-    Parameters
-    ----------
-    filename : str | Path
-        Filename or path of the extracted ultrasound frame.
-    session : Session
-        Session that the frame belongs to.
-    recording : Recording
-        Recording that the frame belongs to.
-    aggregate_meta : AggregateImageParameters
-        The parameters of the AggregateImage to be dumped in a file along with
-        the session and recording information.
-    """
-    if not isinstance(filename, Path):
-        filename = Path(filename)
-    meta_filename = filename.with_suffix('.txt')
-    with meta_filename.open('w', encoding='utf-8') as file:
-        file.write(
-            f"Metadata for AggregateImage extracted by SATKIT to {filename}.\n")
-        file.write(f"Session path: {session.recorded_path}\n")
-        file.write(f"Participant ID: {recording.meta_data.participant_id}\n")
-        file.write(f"Recording filename: {recording.name}\n")
-        file.write(f"Recorded at: {recording.meta_data.time_of_recording}\n")
-        file.write(f"Prompt: {recording.meta_data.prompt}\n")
-
-        nestedtext.dump(aggregate_meta.model_dump(), file,
-                        converters=nested_text_converters)
-        _logger.debug("Wrote file %s.", meta_filename)
-
-
-def export_aggregate_image_and_meta(
-        image: AggregateImage,
-        session: Session,
-        recording: Recording,
-        path: Path
-):
-    if path.is_dir():
-        filename = image.name.replace(" ", "_")
-        filepath = path / filename
-    else:
-        filepath = path
-
-
-    _logger.debug("Wrote file %s.", filepath)
-
-    export_aggregate_image_meta(
-        filename=filepath.with_suffix(".txt"),
-        session=session,
-        recording=recording,
-        aggregate_meta=image.meta_data
-    )
-
